backend/requirements.txt
fastapi
uvicorn[standard]
websockets
anthropic
httpx
python-dotenv
